added file permissions and versioning
the owner of the file can move update, move the file to the previous
versions folder and if the file exists a confirmation is asked to
overwrite the current file.
the owner can also assign a new ownership for the file.
When a user modifies anything in the File Tree, the other users receive
a message to refresh their File Tree.